Fix a deprecation warning and move AkiwrapperImpl to .entities.impl
